requirements.txt
playwright>=1.40
httpx>=0.25
pydantic>=2.0
pyyaml>=6.0
aiohttp>=3.9
winotify>=1.1; sys_platform == "win32"
pystray>=0.19; sys_platform == "win32"
Pillow>=10.0; sys_platform == "win32"
